Moved Melee Attack Logic out of PlayerController to separate MeleeAttack script
Consequently re-serialized player character prefab
