+++ Assets/Prefabs/Warrior_Red.prefab
@@ -12,6 +12,7 @@ GameObject:
   - component: {fileID: 8030250919073706537}
   - component: {fileID: 5565923892229975813}
   - component: {fileID: 6556081880917193565}
+  - component: {fileID: 7072528679712103292}
   - component: {fileID: 571950108984476302}
   - component: {fileID: 484125258941416843}
   m_Layer: 6
@@ -125,8 +126,23 @@ MonoBehaviour:
   playerSprite: {fileID: 8030250919073706537}
   playerAnimator: {fileID: 5565923892229975813}
   playerRigidBody: {fileID: 571950108984476302}
-  weaponCollider: {fileID: 2271233326461272848}
+  meleeAttack: {fileID: 7072528679712103292}
   moveSpeed: 0.05
+--- !u!114 &7072528679712103292
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4439481353513776549}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 30416b861383bcf4da55bcc50917199d, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  playerController: {fileID: 6556081880917193565}
+  playerAnimator: {fileID: 5565923892229975813}
+  weaponCollider: {fileID: 2271233326461272848}
 --- !u!50 &571950108984476302
 Rigidbody2D:
   serializedVersion: 4
