+++ Assets/Scripts/MeleeAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30416b861383bcf4da55bcc50917199d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
