+++ Assets/Scripts/PlayerController.cs
@@ -7,17 +7,13 @@ namespace MustafaNaqvi
         [SerializeField] private SpriteRenderer playerSprite;
         [SerializeField] private Animator playerAnimator;
         [SerializeField] private Rigidbody2D playerRigidBody;
-        [SerializeField] private BoxCollider2D weaponCollider;
+        [SerializeField] private MeleeAttack meleeAttack;
         [SerializeField] private float moveSpeed;
 
         private static readonly int Running = Animator.StringToHash("Running");
-        private static readonly int AttackHorizontal = Animator.StringToHash("Attack_Horizontal");
-        private static readonly int AttackUp = Animator.StringToHash("Attack_Up");
-        private static readonly int AttackDown = Animator.StringToHash("Attack_Down");
 
-        private FacingDirection _facingDirection;
+        internal FacingDirection FacingDirection;
         private float _horizontal, _vertical;
-        private bool _attacking;
 
         private void Start()
         {
@@ -29,6 +25,9 @@ namespace MustafaNaqvi
 
             if (ReferenceEquals(playerRigidBody, null) && TryGetComponent<Rigidbody2D>(out var rigidBody2D))
                 playerRigidBody = rigidBody2D;
+
+            if (ReferenceEquals(meleeAttack, null) && TryGetComponent<MeleeAttack>(out var ma))
+                meleeAttack = ma;
         }
 
         private void Update()
@@ -36,7 +35,6 @@ namespace MustafaNaqvi
             HandleInput();
             HandleMovement();
             HandleRotation();
-            HandleMeleeAttack();
         }
 
         private void HandleInput()
@@ -48,7 +46,8 @@ namespace MustafaNaqvi
         private void HandleMovement()
         {
             var movement = new Vector3(_horizontal, _vertical, 0f).normalized;
-            if (Equals(movement, Vector3.zero) || _attacking)
+            if (ReferenceEquals(meleeAttack, null)) return;
+            if (Equals(movement, Vector3.zero) || meleeAttack.Attacking)
             {
                 playerAnimator.SetBool(Running, false);
                 return;
@@ -62,7 +61,7 @@ namespace MustafaNaqvi
         {
             if (_vertical.Equals(0f) && _horizontal.Equals(0)) return;
 
-            _facingDirection = _vertical switch
+            FacingDirection = _vertical switch
             {
                 0f when _horizontal > 0f => FacingDirection.Right,
                 0f when _horizontal < 0f => FacingDirection.Left,
@@ -75,7 +74,7 @@ namespace MustafaNaqvi
                 _ => FacingDirection.Right
             };
 
-            playerSprite.flipX = _facingDirection switch
+            playerSprite.flipX = FacingDirection switch
             {
                 FacingDirection.Right => false,
                 FacingDirection.Left => true,
@@ -83,89 +82,6 @@ namespace MustafaNaqvi
                 FacingDirection.Down => false,
                 _ => playerSprite.flipX
             };
-        }
-
-        private void HandleMeleeAttack()
-        {
-            if (!Input.GetKeyUp(KeyCode.Space) || _attacking) return;
-
-            _attacking = true;
-
-            switch (_facingDirection)
-            {
-                case FacingDirection.Right or FacingDirection.Left:
-                    PlayAttackAnimation(AttackHorizontal);
-                    break;
-
-                case FacingDirection.Up:
-                    PlayAttackAnimation(AttackUp);
-                    break;
-
-                case FacingDirection.Down:
-                    PlayAttackAnimation(AttackDown);
-                    break;
-            }
-
-            HandleWeaponCollider();
-        }
-
-        // ReSharper disable Unity.PerformanceAnalysis
-        private void PlayAttackAnimation(int id)
-        {
-            playerAnimator.ResetTrigger(AttackHorizontal);
-            playerAnimator.ResetTrigger(AttackUp);
-            playerAnimator.ResetTrigger(AttackDown);
-            playerAnimator.SetTrigger(id);
-
-            Invoke(nameof(ResetAttacking), playerAnimator.GetCurrentAnimatorClipInfo(0)[0].clip.length);
-        }
-
-        private void HandleWeaponCollider()
-        {
-            if (ReferenceEquals(weaponCollider, null)) return;
-
-            var offset = weaponCollider.offset;
-            switch (_facingDirection)
-            {
-                case FacingDirection.Right:
-                    offset.x += 0.5f;
-                    break;
-
-                case FacingDirection.Left:
-                    offset.x -= 0.5f;
-                    break;
-
-                case FacingDirection.Up:
-                    offset.y += 0.5f;
-                    break;
-
-                case FacingDirection.Down:
-                    offset.y -= 0.5f;
-                    break;
-            }
-
-            weaponCollider.offset = offset;
-            weaponCollider.size = Vector2.one * 0.75f;
-        }
-
-        private void ResetAttacking()
-        {
-            _attacking = false;
-            if (ReferenceEquals(weaponCollider, null)) return;
-            weaponCollider.offset = Vector2.zero;
-            weaponCollider.size = Vector2.one * 0.25f;
-        }
-
-        private void GiveDamage()
-        {
-            // Give Damage
-        }
-
-        private void OnCollisionEnter2D(Collision2D other)
-        {
-            if (!other.collider.CompareTag("Damageable")) return;
-            if (!ReferenceEquals(other.otherCollider, weaponCollider)) return;
-            GiveDamage();
         }
     }
